pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aita-core"
version = "0.1.0"
description = "Shared core for AI Teaching Assistant (AITA) chatbots"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.30.0",
    "openai>=1.0.0",
    "faiss-cpu>=1.7.4",
    "pdfminer.six>=20221105",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "google-auth-oauthlib>=1.0.0",
    "PyJWT>=2.0.0",
    "requests>=2.28.0",
    "pandas>=2.0.0",
]

[project.urls]
Repository = "https://github.com/umnCETransportation/aita-core"
